Minor changes.
git-svn-id: http://trypython.googlecode.com/svn/trunk@5 ab4eaae6-8d89-11de-8988-25e8a898b690

+++ maketutorial.py
@@ -24,7 +24,8 @@ for name in input_files:
 handle.close()
 
 doc_dir = os.path.join(this_dir, 'trypython', 'app', 'docs')
-shutil.rmtree(doc_dir)
+if os.path.isdir(doc_dir):
+    shutil.rmtree(doc_dir)
 os.mkdir(doc_dir)
 
 for index, name in enumerate(input_files):
